Move DownloadColumn to generic wagtail.admin.ui.tables
