+++ wagtail/admin/templates/wagtailadmin/tables/download_cell.html
